// app/containers/View/types.ts
export interface IExcuteSqlStatisticsResponse<T> {
  data: T,
  status: number
}
// ...
export interface colBase {
  key: number | string,
  title: string,
  dataIndex: number | string
}
export interface statisticsResponseData {
  col: ColumnProps<colBase>[],
  data: object[]
}